+++ src/main/java/com/prgrms/commandLineApplication/voucher/PercentDiscountVoucher.java
@@ -1,7 +1,5 @@
 package com.prgrms.commandLineApplication.voucher;
 
-import com.prgrms.commandLineApplication.validation.VoucherValidation;
-
 import java.util.UUID;
 
 public class PercentDiscountVoucher extends Voucher {
@@ -13,7 +11,7 @@ public class PercentDiscountVoucher extends Voucher {
   }
 
   public static PercentDiscountVoucher of(UUID voucherId, String voucherType, int discountAmount) {
-    VoucherValidation.checkPercentDiscountAmount(discountAmount);
+    VoucherValidator.checkPercentDiscountAmount(discountAmount);
     return new PercentDiscountVoucher(voucherId, voucherType, discountAmount);
   }
