+++ src/main/java/com/prgrms/commandLineApplication/voucher/Voucher.java
@@ -1,7 +1,5 @@
 package com.prgrms.commandLineApplication.voucher;
 
-import com.prgrms.commandLineApplication.validation.VoucherValidation;
-
 import java.util.UUID;
 
 public abstract class Voucher {
@@ -11,10 +9,22 @@ public abstract class Voucher {
   protected final String voucherType;
 
   protected Voucher(UUID voucherId, String voucherType, int discountAmount) {
-    VoucherValidation.checkId(voucherId);
+    VoucherValidator.checkId(voucherId);
     this.voucherId = voucherId;
     this.voucherType = voucherType;
     this.discountAmount = discountAmount;
+  }
+
+  public UUID getVoucherId() {
+    return voucherId;
+  }
+
+  public int getDiscountAmount() {
+    return discountAmount;
+  }
+
+  public String getVoucherType() {
+    return voucherType;
   }
 
   abstract int discount(int price);
